WIP Offline messages are being retrieved, synchronized, and displayed on demand, however I'm running into a infinite loop glitch, likely something to do with my field map.

+++ app/src/main/java/com/nextcloud/talk/data/database/mappers/ChatMessageMapUtils.kt
@@ -13,7 +13,7 @@ import com.nextcloud.talk.models.json.chat.ChatMessage
 
 fun ChatMessageJson.asEntity() =
     ChatMessageEntity(
-        id = id.toLong(),
+        id = id,
         message = message,
         token = token,
         actorType = actorType,
